+++ lib/js/reference-list/components/Card.tsx
@@ -12,23 +12,10 @@ export class Card extends React.Component<CardProps, {}> {
     }
 
     render() {
-
-        let style: React.CSSProperties = {
-            borderBottom: '1px solid #E5E5E5',
-            padding: 5,
-            cursor: 'pointer',
-        };
-
-        if (this.props.isSelected) {
-            style.backgroundColor = 'rgba(243, 255, 62, 0.2)';
-            style.textShadow = '0px 0px 0.1px';
-        }
-
         return (
             <div
-                className='abt-card'
+                className={this.props.isSelected ? 'abt-card selected' : 'abt-card'}
                 onClick={this.props.onClick}
-                style={style}
                 dangerouslySetInnerHTML={{ __html: this.props.html }} />
         );
     }